requirements.txt
boto3==1.29.3
azure-mgmt-consumption==11.0.0
google-cloud-billing==1.11.0
pymongo==4.6.0
fastapi==0.104.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0